% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment.R
\name{add_enrollment}
\alias{add_enrollment}
\title{Enroll a user into the course}
\usage{
add_enrollment(course_id, user_id, type, state, section = F, ...)
}
\arguments{
\item{course_id}{ID of the course (or section, see below)}

\item{user_id}{ID of the user}

\item{type}{enrollment type}

\item{state}{enrollment state}

\item{section}{if TRUE, course_id should be a section_id instead of a course_id and students will be enrolled into this section}

\item{...}{any other arguments passed to the API}
}
\description{
Enroll a user into the course
}
