% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{add_event}
\alias{add_event}
\title{Add an event to the course calendar}
\usage{
add_event(
  course_id = 16248,
  title = "Murphy's Laws",
  location = "MURPH101",
  repeats = 0,
  frequency = c("weekly", "monthly", "daily"),
  iterator = FALSE,
  day = Sys.Date(),
  start_time = "08:00:00",
  end_time = "10:00:00",
  timezone = "Pacific/Auckland",
  ...
)
}
\arguments{
\item{course_id}{ID of the course}

\item{title}{event title}

\item{location}{location, e.g room}

\item{repeats}{number of repeats or 0 (single event)}

\item{frequency}{if repeats>0, corresponding frequency (otherwise ignored)}

\item{iterator}{boolean, add an event number when there are repeated events}

\item{day}{if repeats>0, corresponding frequency (otherwise ignored)}

\item{start_time}{start_time}

\item{end_time}{end_time}

\item{timezone}{timezone}

\item{...}{any other arguments passed to the API}
}
\description{
See https://canvas.instructure.com/doc/api/calendar_events.html for all options
}
