% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubrics.R
\name{associate_rubric}
\alias{associate_rubric}
\title{Associate a rubric to an assignment}
\usage{
associate_rubric(course_id, rubric_id, association_id)
}
\arguments{
\item{course_id}{Course ID}

\item{rubric_id}{Rubric ID}

\item{association_id}{Assignment ID for the assignment to be associated with the rubric}
}
\value{
(Invisibly) dataframe representation of the RubricAssociation object
}
\description{
TODO: Support all of the options in the API: <https://canvas.instructure.com/doc/api/rubrics.html#method.rubric_associations.create>
}
