% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStudentInCourse.R
\name{checkStudentInCourse}
\alias{checkStudentInCourse}
\title{Title}
\usage{
checkStudentInCourse(course.id = NULL, course.code = NULL, user.id = NULL,
  sis.user.id = NULL, account.id = NULL)
}
\arguments{
\item{course.id}{numeric The Canvas Course ID}

\item{course.code}{character "Vakcode" matches the course code returned by API}

\item{user.id}{numeric The Canvas User ID}

\item{sis.user.id}{numeric The SIS User ID}

\item{account.id}{numeric The Canvas Account ID}
}
\value{
Vector with either true or false, should be same length as input.
}
\description{
Title
}
