% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submissions.R
\name{comment_submission}
\alias{comment_submission}
\title{Comment on a submission}
\usage{
comment_submission(course_id, assignment_id, user_id, comm, to_group = TRUE,
  visible = TRUE)
}
\arguments{
\item{course_id}{A valid canvas course id}

\item{assignment_id}{A valid assignment id}

\item{user_id}{A valid user id}

\item{comm}{What comment to give}

\item{to_group}{Whether the comment should be sent to all members of the group}

\item{visible}{Whether the comment should be visible to student or not}
}
\value{
Invisibly performs the PUT call
}
\description{
See https://canvas.instructure.com/doc/api/submissions.html#method.submissions_api.update
This plays nicely with purrr::map2, so it's easy to assign comments en masse
to students.
}
\examples{
comment_submission(1350207, 5681164, 4928217, "test")
}
