% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_course.R
\name{create_canvas_course}
\alias{create_canvas_course}
\title{Create a course}
\usage{
create_canvas_course(account_id, name = NULL, course_code = NULL,
  start_at = NULL, end_at = NULL, license = NULL, is_public = NULL,
  is_public_to_auth_users = NULL, public_syllabus = NULL,
  public_syllabus_to_auth = NULL, public_description = NULL,
  allow_student_wiki_edits = NULL, allow_wiki_comments = NULL,
  allow_student_forum_attachments = NULL, open_enrollment = NULL,
  self_enrollment = NULL, restrict_enrollments_to_course_dates = NULL,
  term_id = NULL, sis_course_id = NULL, integration_id = NULL,
  hide_final_grades = NULL, apply_assignment_group_weights = NULL,
  time_zone = NULL, offer = NULL, enroll_me = NULL,
  default_view = NULL, syllabus_body = NULL,
  grading_standard_id = NULL, course_format = NULL,
  enable_sis_activation = NULL)
}
\arguments{
\item{account_id}{a valid account id}

\item{name}{The name of the course. If omitted, the course will be named “Unnamed Course.”}

\item{course_code}{The course code for the course.}

\item{start_at}{Course start date in ISO8601 format, e.g. 2011-01-01T01:00Z}

\item{end_at}{Course end date in ISO8601 format, e.g. 2011-01-01T01:00Z}

\item{license}{The name of the licensing. Should be one of the following abbreviations (a descriptive name is included in parenthesis for reference): 'private' (Private Copyrighted)
'cc_by_nc_nd' (CC Attribution Non-Commercial No Derivatives) 'cc_by_nc_sa' (CC Attribution Non-Commercial Share Alike)
'cc_by_nc' (CC Attribution Non-Commercial) 'cc_by_nd' (CC Attribution No Derivatives) 'cc_by_sa' (CC Attribution Share Alike)
'cc_by' (CC Attribution) 'public_domain' (Public Domain).}

\item{is_public}{boolean, Set to true if course is public to both authenticated and unauthenticated users.}

\item{is_public_to_auth_users}{boolean, Set to true if course is public only to authenticated users.}

\item{public_syllabus}{boolean, Set to true to make the course syllabus public.}

\item{public_syllabus_to_auth}{boolean, Set to true to make the course syllabus public for authenticated users.}

\item{public_description}{A publicly visible description of the course.}

\item{allow_student_wiki_edits}{boolean, If true, students will be able to modify the course wiki.}

\item{allow_wiki_comments}{boolean, If true, course members will be able to comment on wiki pages.}

\item{allow_student_forum_attachments}{boolean, If true, students can attach files to forum posts.}

\item{open_enrollment}{boolean, Set to true if the course is open enrollment.}

\item{self_enrollment}{boolean, Set to true if the course is self enrollment.}

\item{restrict_enrollments_to_course_dates}{boolean, Set to true to restrict user enrollments to the start and end dates of the course.}

\item{term_id}{The unique ID of the term to create to course in.}

\item{sis_course_id}{The unique SIS identifier.}

\item{integration_id}{The unique Integration identifier.}

\item{hide_final_grades}{boolean, If this option is set to true, the totals in student grades summary will be hidden.}

\item{apply_assignment_group_weights}{boolean, Set to true to weight final grade based on assignment groups percentages.}

\item{time_zone}{The time zone for the course. Allowed time zones are IANA time zones or friendlier Ruby on Rails time zones.}

\item{offer}{boolean, If this option is set to true, the course will be available to students immediately.}

\item{enroll_me}{boolean, Set to true to enroll the current user as the teacher.}

\item{default_view}{The type of page that users will see when they first visit the course. Options: 'feed', 'wiki', 'modules', syllabus', 'assignments'}

\item{syllabus_body}{The syllabus body for the course}

\item{grading_standard_id}{The grading standard id to set for the course. If no value is provided for this argument the current grading_standard will be un-set from this course.}

\item{course_format}{Optional. Specifies the format of the course. (Should be 'on_campus', 'online', or 'blended')}

\item{enable_sis_activation}{When true, will first try to re-activate a deleted course with matching sis_course_id if possible.}
}
\value{
invisible
}
\description{
Create a course
}
\examples{
\dontrun{create_canvas_course(account_id = 32243, name = "Soren Kierkegaard")}
}
