% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploads.R
\name{create_canvas_module}
\alias{create_canvas_module}
\title{Create a course module}
\usage{
create_canvas_module(
  course_id,
  name,
  unlock_at = NULL,
  position = NULL,
  require_sequential_progress = NULL,
  prerequisite_module_ids = NULL,
  publish_final_grade = NULL
)
}
\arguments{
\item{course_id}{a valid course id}

\item{name}{the module name (only parameter required)}

\item{unlock_at}{datetime - The day/time the module is unlocked. Accepts times in ISO 8601 format, e.g. 2014-10-21T18:48:00Z.}

\item{position}{integer - The position of this module.}

\item{require_sequential_progress}{boolean}

\item{prerequisite_module_ids}{integer}

\item{publish_final_grade}{boolean - whether the final grade should be published upon completion}
}
\value{
invisible
}
\description{
Create a course module
}
\examples{
create_canvas_module(course_id = 432432, name = "Module 1")

}
