% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploads.R
\name{create_cavnas_module_item}
\alias{create_cavnas_module_item}
\title{Create a module item}
\usage{
create_cavnas_module_item(
  course_id,
  module_id,
  type,
  content_id = NULL,
  title = NULL,
  position = NULL,
  indent = NULL,
  page_url = NULL,
  external_url = NULL,
  new_tab = NULL
)
}
\arguments{
\item{course_id}{a valid course id}

\item{module_id}{a valid module id}

\item{type}{string - type of module item (must be one of: File, Page, Discussion, Assignment, Quiz, SubHeader, ExternalUrl, ExternalTool)}

\item{content_id}{integer - The id of the content to link to the module item. Required, except for ‘ExternalUrl’, ‘Page’, and ‘SubHeader’ types.}

\item{title}{integer - The name of the module item and associated content}

\item{position}{integer - The position of this item in the module}

\item{indent}{integer - 0-based indent level; module items may be indented to show a hierarchy}

\item{page_url}{string - Suffix for the linked wiki page (e.g. ‘front-page’). Required for ‘Page’ type.}

\item{external_url}{string - External url that the item points to. (Required for ‘ExternalUrl’ and ‘ExternalTool’ types)}

\item{new_tab}{boolean - Whether the external tool opens in a new tab. Only applies to ‘ExternalTool’ type.}
}
\value{
invisible
}
\description{
Create a module item
}
\examples{
create_canvas_module_item(
  course_id = 432432,
  module_id = 1,
  type = "SubHeader",
  title = "Before Class")
}
