% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_migrations.R
\name{create_course_content_migration}
\alias{create_course_content_migration}
\title{Create a course content migration}
\usage{
create_course_content_migration(course_id,
  migration_type = "course_copy_importer", name = NULL,
  file_url = NULL, source_course_id = NULL, folder_id = NULL,
  overwrite_quizzes = NULL, question_bank_id = NULL,
  question_bank_name = NULL, shift_dates = NULL,
  old_start_date = NULL, old_end_date = NULL, new_start_date = NULL,
  new_end_date = NULL, day_substitutions = NULL, remove_dates = NULL)
}
\arguments{
\item{course_id}{a valid course id (the course to copy content into)}

\item{migration_type}{Migration type. Default allowed values: canvas_cartridge_importer, common_cartridge_importer,
course_copy_importer, zip_file_importer, qti_converter, moodle_converter}

\item{name}{Required if uploading a file. This is the first step in uploading a file to the content migration.}

\item{file_url}{A URL to download the file from. Must not require authentication.}

\item{source_course_id}{a valid course id (the course to copy content from)}

\item{folder_id}{The folder to unzip the .zip file into for a zip_file_import.}

\item{overwrite_quizzes}{boolean, Whether to overwrite quizzes with the same identifiers between content packages.}

\item{question_bank_id}{The existing question bank ID to import questions into if not specified in the content package.}

\item{question_bank_name}{The question bank to import questions into if not specified in the content package, if both bank id and name are set, id will take precedence.}

\item{shift_dates}{boolean, Whether to shift dates in the copied course}

\item{old_start_date}{The original start date of the source content/course}

\item{old_end_date}{The original end date of the source content/course}

\item{new_start_date}{The new start date for the content/course}

\item{new_end_date}{The new end date for the content/course}

\item{day_substitutions}{Move anything scheduled for day 'X' to the specified day. (0-Sunday, 1-Monday, 2-Tuesday, 3-Wednesday, 4-Thursday, 5-Friday, 6-Saturday)}

\item{remove_dates}{boolean, Whether to remove dates in the copied course. Cannot be used in conjunction with shift_dates.}
}
\value{
invisible
}
\description{
Create a course content migration
}
\examples{
\dontrun{create_course_content_migration(course_id = 1532751, source_course_id = 1532594)}
}
