% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploads.R
\name{create_course_folder}
\alias{create_course_folder}
\title{Create a course folder}
\usage{
create_course_folder(
  course_id,
  name,
  parent_folder_id = NULL,
  parent_folder_path = ""
)
}
\arguments{
\item{course_id}{a valid course id}

\item{name}{name of the folder (required)}

\item{parent_folder_id}{The id of the folder to store the file in. If this and parent_folder_path
are sent an error will be returned. If neither is given, a default folder will be used.}

\item{parent_folder_path}{The path of the folder to store the new folder in. If this and
parent_folder_id are sent an error will be returned. If neither is given, a default folder will be used.}
}
\value{
invisible
}
\description{
Create a course folder
}
\examples{
create_course_folder(34232, name = "activities")
}
