% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{create_wpage}
\alias{create_wpage}
\title{Create page in course}
\usage{
create_wpage(course_id, title, body, editing_roles = "teachers",
  published = FALSE)
}
\arguments{
\item{course_id}{a valid course id}

\item{title}{a string.  The title for the new page.}

\item{body}{a string    The content for the new page.}

\item{editing_roles}{a string.  Which user roles are allowed to edit this page. Any combination of these roles is allowed (separated by commas).  Allowed values: teachers, students, members, public}

\item{published}{a boolean.    Whether the page is published (true) or draft state (false).}
}
\value{
empty
}
\description{
Create page in course
}
