% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz-questions.R
\name{delete_quiz_question}
\alias{delete_quiz_question}
\title{Delete a question from a classic quiz}
\usage{
delete_quiz_question(course_id, quiz_id, question_id)
}
\arguments{
\item{course_id}{A valid course id.}

\item{quiz_id}{A valid classic quiz id.}

\item{question_id}{A valid quiz question id.}
}
\value{
The httr response, invisibly.
}
\description{
Delete a question from a classic quiz
}
\examples{
\dontrun{delete_quiz_question(20, 123, 456)}
}
