% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizes.R
\name{delete_quiz_report}
\alias{delete_quiz_report}
\title{Delete quiz report}
\usage{
delete_quiz_report(course_id, quiz_id, report_id)
}
\arguments{
\item{course_id}{a valid course id}

\item{quiz_id}{a valid quiz id.}

\item{report_id}{a valid report id.}
}
\description{
Delete quiz report
}
