% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{delete_wpage}
\alias{delete_wpage}
\title{Delete page from course}
\usage{
delete_wpage(course_id, page_url)
}
\arguments{
\item{course_id}{a valid course id}

\item{page_url}{a valid page url.}
}
\value{
invisible
}
\description{
Delete page from course
}
