% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_query}
\alias{do_query}
\title{Execute a query on the remove API}
\usage{
do_query(
  endpoint,
  args = NULL,
  method = "GET",
  process_response = (method == "GET")
)
}
\arguments{
\item{endpoint}{the API endpoint to call, with or without the canvas domain. You can give a vector of parts which will be joined with slashes.}

\item{args}{a list of arguments for the call}

\item{method}{GET or POST}

\item{process_response}{if TRUE (default for GET requests), paginate results and return a data frame}
}
\value{
A data.frame if process_response is TRUE, otherwise an httr response
}
\description{
This function allows you to call methods which are not specifically exposed by this API yet
}
\examples{
# A get request to the announcements endpoint (replicating get_announcements):
do_query("announcements", list(`context_codes[]`="course_1234"))

# A post request to the group membership endpoint (replicating add_group_user):
do_query(c("groups", 123, "memberships"), list(user_id=1), method = "POST")
}
