% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{duplicate_wpage}
\alias{duplicate_wpage}
\title{Duplicate course page}
\usage{
duplicate_wpage(course_id, page_url)
}
\arguments{
\item{course_id}{a valid course id}

\item{page_url}{name of the page url. Just the final portion.}
}
\value{
invisible
}
\description{
Duplicate course page
}
