% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment.R
\name{add_enrollments}
\alias{add_enrollments}
\title{Enroll user(s) into a course (or multiple courses)}
\usage{
add_enrollments(course_id, user_ids, type = c("StudentEnrollment",
  "TeacherEnrollment", "TaEnrollment", "ObserverEnrollment",
  "DesignerEnrollment"), state = c("invited", "active", "inactive"),
  section = F, ...)
}
\arguments{
\item{course_id}{ID of the course (or section, see below)}

\item{type}{enrollment type}

\item{state}{enrollment state}

\item{section}{if TRUE, course_id should be a section_id instead of a course_id and students will be enrolled into this section}

\item{...}{any other arguments passed to the API}

\item{user_id}{ID of the user}
}
\description{
Enrolls the given user(s) into the course. If multiple course_ids are given, it should be of the same length as user_id,
and each user will be enrolled into the corresponding course.
}
