% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-data.R
\name{get_account_course_list}
\alias{get_account_course_list}
\title{Function to list courses for a specific account or all courses.}
\usage{
get_account_course_list(acc_id = NULL, include = NULL)
}
\arguments{
\item{acc_id}{Optional argument to specify courses for a specific account id}

\item{include}{Optional argument to specify additional information}
}
\value{
data frame
}
\description{
Function to list courses for a specific account or all courses.
}
\examples{
#' get_account_course_list()
#' get_account_course_list(acc_id = 123)
}
