% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{get_assignment_list}
\alias{get_assignment_list}
\title{Function to list all assignments.}
\usage{
get_assignment_list(course_id = NULL)
}
\arguments{
\item{course_id}{Course ID}
}
\value{
data frame
}
\description{
Function to list all assignments.
}
\examples{
#' get_assignment_list()
}
