% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-data.R
\name{get_course_items}
\alias{get_course_items}
\title{Get course items}
\usage{
get_course_items(course_id, item = NULL)
}
\arguments{
\item{course_id}{valid Canvas course id}

\item{item}{one of "settings", "discussion_topics", "todo", "enrollments", "features", "files", "modules", "front_page", "pages", "quizzes"}
}
\value{
data.frame
}
\description{
Get course items
}
\examples{
#' get_course_items(course_id = 20, item = "settings")
#' get_course_items(course_id = 20, item = "enrollments")
}

