% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-data.R
\name{get_course_list}
\alias{get_course_list}
\title{Function to list all courses.}
\usage{
get_course_list(user_id = NULL)
}
\arguments{
\item{user_id}{Optional argument to specify courses for specific user id}
}
\value{
data.frame
}
\description{
Function to list all courses.
}
\examples{
#' get_course_list()
#' get_course_list(user_id = 366)
}

