% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-data.R
\name{get_course_permissions}
\alias{get_course_permissions}
\title{Get course permissions for the authenticated user}
\usage{
get_course_permissions(course_id, permissions = NULL)
}
\arguments{
\item{course_id}{The ID of the course for which you want to get permission information.}

\item{permissions}{A vector of permission names to check against the authenticated user.
Permission names are documented in the Create a role endpoint.}
}
\value{
A named list with permission names and their respective boolean values (TRUE/FALSE).
}
\description{
This function returns permission information for the calling user in the given course.
}
