% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{get_course_user_groups}
\alias{get_course_user_groups}
\title{Get all users in a course and which group they are signed up for}
\usage{
get_course_user_groups(course_id)
}
\arguments{
\item{course_id}{which course}
}
\value{
dataframe with user name, user id, and group id
}
\description{
Get all users in a course and which group they are signed up for
}
