% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discussions.R
\name{get_discussion_entries}
\alias{get_discussion_entries}
\title{Get discussion entries by user}
\usage{
get_discussion_entries(course_id, topic_id)
}
\arguments{
\item{course_id}{The course ID.}

\item{user_id}{The user ID.}
}
\value{
A data frame containing discussion entries for the specified user.
}
\description{
Get discussion entries by user
}
