% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discussions.R
\name{get_discussions_context}
\alias{get_discussions_context}
\alias{get_discussion_id}
\alias{update_discussion_id}
\title{Discussion functions}
\usage{
get_discussions_context(object_id, object_type = "courses",
  include = NULL)

get_discussion_id(discussion_id, object_id, object_type = "courses")

update_discussion_id(discussion_id, object_id, message,
  object_type = "courses")
}
\arguments{
\item{object_id}{course or group id}

\item{object_type}{"courses" or "groups"}

\item{include}{If "all_dates" is passed, all dates associated with graded discussions' assignments will be included.}

\item{discussion_id}{specific id of discussion to get/update}

\item{message}{new body of discussion id}
}
\value{
discussions belonging to requested context

single discussion

silently sends put request and updates
}
\description{
A group of functions which deal with \href{https://canvas.instructure.com/doc/api/discussion_topics.html}{discussion topics}.
}
\details{
\itemize{
\item \code{get_discussions_context}: get all discussions belonging to a course or group. Note that theoretically this should include announcements, as they are technically discussions, but does not. Use \code{get_announcements} instead.
}
}
\examples{
get_discussions_context(4371405)
get_discussion_id(4371405, 1350207)
update_discussion_id(4371405, 1350207, newtext)
}
