% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploads.R
\name{get_folder_list}
\alias{get_folder_list}
\title{Function to list all folders.}
\usage{
get_folder_list(course_id)
}
\arguments{
\item{course_id}{Course ID}
}
\value{
data frame
}
\description{
Function to list all folders.
}
\examples{
get_folder_list(1234)
}
