% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-data.R
\name{get_groups}
\alias{get_groups}
\title{Get groups}
\usage{
get_groups(course_id)
}
\arguments{
\item{course_id}{which course}
}
\value{
returns group data for a given course
}
\description{
Get groups
}

