% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{get_groups_context}
\alias{get_groups_context}
\title{Get groups given context}
\usage{
get_groups_context(object_id, object_type = "courses")
}
\arguments{
\item{object_id}{id for given type}

\item{object_type}{course or account}
}
\value{
returns groups available for a given course or account
}
\description{
Returns the list of active groups in the given context that are visible to user.
}
