% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{get_module_list}
\alias{get_module_list}
\title{Get List of modules in a course}
\usage{
get_module_list(course_id = NULL)
}
\arguments{
\item{course_id}{a valid course id}
}
\value{
dataframe of modules
}
\description{
Get List of modules in a course
}
\examples{
get_module_list(course_id = 27)
}
