% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-data.R
\name{get_outcome_results}
\alias{get_outcome_results}
\title{Function to return course outcome results.}
\usage{
get_outcome_results(course_id)
}
\arguments{
\item{course_id}{A valid Canvas course id}
}
\value{
data frame
}
\description{
Returns a data.frame containing course outcome results.
}
\examples{
#' get_outcome_results(course_id = 20)
}
