% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizes.R
\name{get_quiz}
\alias{get_quiz}
\title{Get data frame of a quiz in course}
\usage{
get_quiz(course_id, quiz_id)
}
\arguments{
\item{course_id}{a valid course id}

\item{quiz_id}{a valid quiz id.}
}
\value{
a tibble with variables specified for the quiz
}
\description{
Get data frame of a quiz in course
}
