% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizes.R
\name{get_quiz_list}
\alias{get_quiz_list}
\title{Get data frame of quizes in course}
\usage{
get_quiz_list(course_id, search = "")
}
\arguments{
\item{course_id}{a valid course id}

\item{search}{a string. The partial title of the pages to match and return. Defaults to `""` or a blank search}
}
\value{
a data.frame with all pages from course.
}
\description{
Get data frame of quizes in course
}
