% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz-questions.R
\name{get_quiz_questions}
\alias{get_quiz_questions}
\title{List questions in a classic quiz}
\usage{
get_quiz_questions(course_id, quiz_id)
}
\arguments{
\item{course_id}{A valid course id.}

\item{quiz_id}{A valid classic quiz id.}
}
\value{
A data frame of quiz questions.
}
\description{
List questions in a classic quiz
}
\examples{
\dontrun{get_quiz_questions(20, 123)}
}
