% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizes.R
\name{get_quiz_report}
\alias{get_quiz_report}
\title{Get data frame of a quiz reports for a quiz in course}
\usage{
get_quiz_report(course_id, quiz_id, report_id, include = "file")
}
\arguments{
\item{course_id}{a valid course id}

\item{quiz_id}{a valid quiz id.}

\item{report_id}{a valid report id.}

\item{include}{Type is string. Whether the output should include documents for the file and/or progress objects associated with this report. (Note: JSON-API only). Allowed values are `file` and `progress`.}
}
\value{
a tibble with variables specified for the quiz
}
\description{
Get data frame of a quiz reports for a quiz in course
}
