% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizes.R
\name{get_quiz_reports}
\alias{get_quiz_reports}
\title{Get data frame of a quiz reports for a quiz in course}
\usage{
get_quiz_reports(course_id, quiz_id, include_all = TRUE)
}
\arguments{
\item{course_id}{a valid course id}

\item{quiz_id}{a valid quiz id.}

\item{include_all}{(boolean) Whether to retrieve reports that consider all the submissions or only the most recent. Defaults to false, ignored for item_analysis reports.}
}
\value{
a tibble with variables specified for the quiz
}
\description{
Get data frame of a quiz reports for a quiz in course
}
