% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubrics.R
\name{get_rubrics}
\alias{get_rubrics}
\title{Rubric functions
Functions for working with rubrics}
\usage{
get_rubrics(course_id)
}
\arguments{
\item{course_id}{A valid canvas course id}
}
\value{
A dataframe of rubrics
}
\description{
Rubric functions
Functions for working with rubrics
}
