% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{get_single_assignment}
\alias{get_single_assignment}
\title{Get a single assignment}
\usage{
get_single_assignment(
  course_id,
  assignment_id,
  include = NULL,
  override_assignment_dates = TRUE,
  needs_grading_count_by_section = FALSE,
  all_dates = FALSE
)
}
\arguments{
\item{course_id}{integer, the ID of the course}

\item{assignment_id}{integer, the ID of the assignment}

\item{include}{string vector, the associations to include with the assignment}

\item{override_assignment_dates}{boolean, whether to apply assignment overrides to the assignment}

\item{needs_grading_count_by_section}{boolean, whether to split up "needs_grading_count" by sections}

\item{all_dates}{boolean, whether to include all dates associated with the assignment}
}
\value{
a data frame with the details of the single assignment
}
\description{
Get a single assignment
}
