% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submissions.R
\name{get_submissions}
\alias{get_submissions}
\title{Get quiz or assignment submissions}
\usage{
get_submissions(course_id, type, type_id)
}
\arguments{
\item{course_id}{A valid canvas course id}

\item{type}{Either "quizzes" or "assignments"}

\item{type_id}{A valid quiz or assignment id}
}
\value{
Student submissions
}
\examples{
#'get_submissions(27, "quizzes", 2915)
#'get_submissions(27, "assignments, 254)
}

