% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-data.R
\name{get_user_items}
\alias{get_user_items}
\title{Get various user items}
\usage{
get_user_items(user_id, item)
}
\arguments{
\item{user_id}{A valid canvas user id}

\item{item}{One of "missing_submissions", "details", "profile", "page_views", "colors", or "avatars"}
}
\value{
data frame
}
\description{
Get various user items
}
\examples{

get_user_items(365, "details")
}
