% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{get_wpage}
\alias{get_wpage}
\title{Get data frame of a page in course}
\usage{
get_wpage(course_id, page_url)
}
\arguments{
\item{course_id}{a valid course id}

\item{page_url}{a valid page url.}
}
\value{
a data.frame with all page from course.
}
\description{
Get data frame of a page in course
}
