% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submissions.R
\name{grade_submission}
\alias{grade_submission}
\title{Grade a submission}
\usage{
grade_submission(course_id, assignment_id, user_id, grade)
}
\arguments{
\item{course_id}{A valid canvas course id}

\item{assignment_id}{A valid assignment id}

\item{user_id}{User's id of submission to grade}

\item{grade}{grade to give}
}
\value{
Invisibly performs the PUT call
}
\description{
This plays nicely with purrr::map2, so it's easy to assign grades en masse
to students.
}
\examples{
grade_submission(1350207, 5681164, 4928217, 80)
}
