% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{list_modules}
\alias{list_modules}
\title{Retrieve a list of modules for a specific course}
\usage{
list_modules(course_id)
}
\arguments{
\item{course_id}{A valid course ID.}
}
\value{
A list containing the modules associated with the specified course.
}
\description{
Retrieve a list of modules for a specific course
}
\examples{
\dontrun{modules <- list_modules(12345)}
}
