% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_response.R
\name{paginate}
\alias{paginate}
\title{Get responses from Canvas API pages}
\usage{
paginate(x, showProgress = T)
}
\arguments{
\item{x}{a httr response object}

\item{showProgress}{if TRUE (default), show a textual progress bar}
}
\value{
unparsed responses
}
\description{
The Canvas headers include a link object (usually), in form:
\code{Link:
<https://canvas.url/api/v1/[...]?page=1&per_page=10>; rel="current",
<https://canvas.url/api/v1/[...]?page=2&per_page=10>; rel="next",
<https://canvas.url/api/v1/[...]?page=1&per_page=10>; rel="first",
<https://canvas.url/api/v1/[...]?page=15&per_page=10>; rel="last"}

In this case, we need to download every page from 1 to 15 to capture all data.
This function parses the response object intelligently, using only HEAD
requests, to figure out these page requirements.
}
\examples{
\dontrun{resp <- canvas_query(url, args, "HEAD")
get_pages(resp)}
}
