% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-data.R
\name{search_courses}
\alias{search_courses}
\title{Search all courses}
\usage{
search_courses(search = NULL)
}
\arguments{
\item{search}{optional search keyword}
}
\value{
data frame
}
\description{
Returns a data.frame of all public courses (optoinally matching the search term)
}
\examples{
  \dontrun{search_courses()}
  \dontrun{search_courses(search="big data")}
}
