% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_canvas_domain}
\alias{set_canvas_domain}
\title{Function to set Canvas domain url}
\usage{
set_canvas_domain(domain)
}
\arguments{
\item{domain}{Canvas domain}
}
\value{
nothing
}
\description{
Given a Canvas domain url, this function adds it to R's
environment variables so it can be found by rcanvas.
}
\examples{
set_canvas_domain("https://canvas.upenn.edu")
}

