% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_canvas_token}
\alias{set_canvas_token}
\title{Function to set Canvas API token}
\usage{
set_canvas_token(token)
}
\arguments{
\item{token}{your API token}
}
\value{
nothing
}
\description{
Given a Canvas token string, this function adds it to R's
environment variables so it can be found by rcanvas.
}
\examples{
set_canvas_token("abc123")
}
