% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{show_wpage_front}
\alias{show_wpage_front}
\title{Show Front Page}
\usage{
show_wpage_front(course_id)
}
\arguments{
\item{course_id}{a valid course id}
}
\value{
data.frame with front page information
}
\description{
Show Front Page
}
\examples{
show_page_front(34232)
}
