% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions.R
\name{submission_response}
\alias{submission_response}
\alias{comment_submission}
\alias{grade_submission}
\title{Respond to submission}
\usage{
comment_submission(course_id, assignment_id, user_id, comm,
  to_group = TRUE, visible = TRUE)

grade_submission(course_id, assignment_id, user_id, grade)
}
\arguments{
\item{course_id}{A valid canvas course id}

\item{assignment_id}{A valid assignment id}

\item{user_id}{A valid user id}

\item{comm}{What comment to give}

\item{to_group}{Whether the comment should be sent to all members of the group}

\item{visible}{Whether the comment should be visible to student or not}

\item{grade}{grade to give}
}
\value{
Invisibly performs the PUT call
}
\description{
See \href{https://canvas.instructure.com/doc/api/submissions.html#method.submissions_api.update}{Canvas API Documentation}.
}
\details{
This plays nicely with \code{purrr::map2}, so it's easy to assign comments or grades
en masse to students.
}
\examples{
\dontrun{comment_submission(1350207, 5681164, 4928217, "test")}
\dontrun{grade_submission(1350207, 5681164, 4928217, 80)}
}
