% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions.R
\name{submissions}
\alias{submissions}
\alias{get_submissions}
\alias{get_submission_single}
\title{Submissions}
\usage{
get_submissions(course_id, type, type_id)

get_submission_single(course_id, type, type_id, user_id)
}
\arguments{
\item{course_id}{A valid canvas course id}

\item{type}{Either "quizzes" or "assignments"}

\item{type_id}{A valid quiz or assignment id}

\item{user_id}{id of user whose submission should be retrieved}
}
\value{
Student submissions
}
\description{
Submissions

\code{get_submissions}: For given course and assignment/quiz, get submissions

`get_submission_single`: Get a single submission, based on user id.
}
\examples{
\dontrun{get_submissions(27, "quizzes", 2915)}
\dontrun{get_submissions(27, "assignments", 254)}
get_submission_single(1350207, "assignments", 5681164, 4928217)
}
