% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{submit_file_upload_assignment}
\alias{submit_file_upload_assignment}
\title{Upload a file to an assignment and submit it}
\usage{
submit_file_upload_assignment(
  course_id,
  assignment_id,
  user_id,
  file_name,
  parent_folder_id = NULL,
  parent_folder_path = "/",
  on_duplicate = "overwrite"
)
}
\arguments{
\item{course_id}{Course ID}

\item{assignment_id}{Assignment ID}

\item{user_id}{User ID}

\item{file_name}{Name of file to be submitted}
}
\value{
data frame
}
\description{
Upload a file to an assignment and submit it
}
