% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_canvas_api}
\alias{test_canvas_api}
\title{Test Canvas API access}
\usage{
test_canvas_api(course_id)
}
\arguments{
\item{course_id}{The course ID for which to retrieve the information.}
}
\value{
A parsed JSON object containing the course information.
}
\description{
This function tests the Canvas API access by retrieving the course information
for the specified course ID.
}
\examples{
course_info <- test_canvas_api(12345)
print(course_info)
}
