% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{update_syllabus_body}
\alias{update_syllabus_body}
\title{Update syllabus body}
\usage{
update_syllabus_body(course_id, body)
}
\arguments{
\item{course_id}{a valid course id}
}
\value{
invisible
}
\description{
Update syllabus body
}
