% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{update_wpage}
\alias{update_wpage}
\title{Update/Create page in course}
\usage{
update_wpage(course_id, page_url, title = NULL, body = NULL,
  editing_roles = "teachers", published = FALSE, notify = FALSE)
}
\arguments{
\item{course_id}{a valid course id}

\item{page_url}{a valid page url.}

\item{title}{a string.  The title for the new page. NOTE: changing a page's title will change its url. The updated url will be returned in the result.}

\item{body}{a string    The content for the new page.}

\item{editing_roles}{a string.  Which user roles are allowed to edit this page. Any combination of these roles is allowed (separated by commas).  Allowed values: teachers, students, members, public}

\item{published}{a boolean.    Whether the page is published (true) or draft state (false).}

\item{notify}{a boolean. Whether participants should be notified when this page changes.}
}
\value{
empty
}
\description{
Update/Create page in course
}
