% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{upload_file}
\alias{upload_file}
\title{upload a file to a folder}
\usage{
upload_file(folder_id)
}
\arguments{
\item{folder_id}{can be one of course, section, or user id.  Can be input as numbers or characters. If a user has multiple enrollments in a context (e.g. as a teacher and a student or in multiple course sections), each enrollment will be listed separately.}
}
\description{
upload a file to a folder
}
