% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apihelpers}
\alias{apihelpers}
\alias{set_canvas_token}
\alias{set_canvas_domain}
\alias{check_token}
\alias{canvas_url}
\alias{make_canvas_url}
\alias{canvas_query}
\alias{iter_args_list}
\alias{sc}
\alias{convert_dates}
\title{Canvas API helpers}
\usage{
set_canvas_token(token)

set_canvas_domain(domain)

check_token()

canvas_url()

make_canvas_url(...)

canvas_query(urlx, args = NULL, type = "GET")

iter_args_list(x, label)

sc(x)

convert_dates(base_date = Sys.Date(), days)
}
\arguments{
\item{token}{your API token}

\item{domain}{Canvas domain}
}
\description{
These functions set your Canvas API token, as well as the Canvas base URL.
These functions are necessary for \code{rcanvas} to run.
}
\examples{
\dontrun{set_canvas_token("abc123")}
\dontrun{set_canvas_domain("https://canvas.upenn.edu")}
}
