% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{apihelpers}
\alias{apihelpers}
\alias{cdenv}
\alias{set_canvas_token}
\alias{set_canvas_domain}
\alias{check_token}
\title{Canvas API helpers}
\format{
An object of class \code{environment} of length 0.
}
\usage{
cdenv

set_canvas_token(token)

set_canvas_domain(domain)

check_token()
}
\arguments{
\item{token}{your API token}

\item{domain}{Canvas domain}
}
\description{
These functions set your Canvas API token, as well as the Canvas base URL.
These functions are necessary for \code{rcanvas} to run.
}
\examples{
set_canvas_token("abc123")
set_canvas_domain("https://canvas.upenn.edu")
}
\keyword{datasets}
