% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploads.R
\name{create_quiz_question}
\alias{create_quiz_question}
\title{Add a question to an existing quiz}
\usage{
create_quiz_question(
  course_id,
  quiz_id,
  question_name = NULL,
  question_text = NULL,
  quiz_group_id = NULL,
  question_type = NULL,
  position = NULL,
  points_possible = NULL,
  correct_comments = NULL,
  incorrect_comments = NULL,
  neutral_comments = NULL,
  text_after_answers = NULL,
  answers = NULL
)
}
\arguments{
\item{course_id}{Course ID}

\item{quiz_id}{Quiz ID}

\item{question_name}{The name of the question.}

\item{question_text}{The text of the question.}

\item{quiz_group_id}{The id of the quiz group to assign the question to.}

\item{question_type}{The type of question. Multiple optional fields depend upon the type of question to be used. Options: calculated_question, essay_question, file_upload_question, fill_in_multiple_blanks_question, matching_question, multiple_answers_question, multiple_choice_question, multiple_dropdowns_question, numerical_question, short_answer_question, text_only_question, true_false_question}

\item{position}{The order in which the question will be displayed in the quiz in relation to other questions.}

\item{points_possible}{The maximum amount of points received for answering this question correctly.}

\item{correct_comments}{The comment to display if the student answers the question correctly.}

\item{incorrect_comments}{The comment to display if the student answers incorrectly.}

\item{neutral_comments}{The comment to display regardless of how the student answered.}

\item{text_after_answers}{no description}

\item{answers}{no description (Answer object)}
}
\value{
Dataframe representation of the question
}
\description{
Add a question to an existing quiz
}
