% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announcements.R
\name{get_announcements}
\alias{get_announcements}
\title{Get announcements for a course}
\usage{
get_announcements(course_id)
}
\arguments{
\item{course_id}{numeric course id}
}
\value{
announcements for a given course
}
\description{
We transform the course id from a numeric (e.g. 123) to a context id
(e.g. course_123).
}
