% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announcements.R
\name{get_announcements}
\alias{get_announcements}
\title{Get announcements for a course}
\usage{
get_announcements(course_id, start_date = NULL, end_date = NULL,
  active_only = FALSE)
}
\arguments{
\item{course_id}{numeric course id}

\item{start_date}{Only return announcements posted since the start_date (inclusive). Defaults to 14 days ago. The value should be formatted as: yyyy-mm-dd or ISO 8601 YYYY-MM-DDTHH:MM:SSZ.}

\item{end_date}{Only return announcements posted before the end_date (inclusive). Defaults to 28 days from start_date. The value should be formatted as: yyyy-mm-dd or ISO 8601 YYYY-MM-DDTHH:MM:SSZ. Announcements scheduled for future posting will only be returned to course administrators.}

\item{active_only}{Only return active announcements that have been published. Applies only to requesting users that have permission to view unpublished items. Defaults to false for users with access to view unpublished items, otherwise true and unmodifiable.}
}
\value{
announcements for a given course
}
\description{
We transform the course id from a numeric (e.g. 123) to a context id
(e.g. course_123).
}
