% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-data.R
\name{get_course_analytics_data}
\alias{get_course_analytics_data}
\title{Function to return course analytics data.}
\usage{
get_course_analytics_data(course_id, type = "assignments",
  user_id = NULL)
}
\arguments{
\item{course_id}{A valid Canvas course id}

\item{type}{One of "assignments", "activity", or "student_summaries"}

\item{user_id}{Optional argument to specify type analytics for individual user id}
}
\value{
data frame or list if user_id is specified
}
\description{
Returns a data.frame of course analytics data. Note: if an individual's user_id is specified,
the function will return a list.
}
\examples{
#' get_course_analytics_data(course_id = 20)
#' get_course_analytics_data(course_id = 17, type = "activity")
#' get_course_analytics_data(course_id = 17, type = "student_summaries", user_id = 366)
}
