% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_gradebook.R
\name{get_course_gradebook}
\alias{get_course_gradebook}
\title{Get a course gradebook}
\usage{
get_course_gradebook(course_id)
}
\arguments{
\item{course_id}{A valid course id}
}
\value{
A gradebook
}
\examples{
get_course_gradebook(20)
}
