% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_gradebook.R
\name{get_course_gradebook}
\alias{get_course_gradebook}
\title{Get a course gradebook}
\usage{
get_course_gradebook(course_id, progress = FALSE)
}
\arguments{
\item{course_id}{A valid course id}

\item{progress}{logical; print page numbers as we go? (Defaults to FALSE)}
}
\value{
A gradebook (in long format)
}
\description{
Get a course gradebook
}
\examples{
get_course_gradebook(20)

}
