% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-data.R
\name{get_course_items}
\alias{get_course_items}
\title{Function to return various course items. See "item" argument below. Omitting the "item argument
returns a course object.}
\usage{
get_course_items(course_id, item, include = NULL)
}
\arguments{
\item{course_id}{valid Canvas course id}

\item{item}{Optional -- one of "settings", "discussion_topics", "todo", "enrollments", "features", "files", "modules", "front_page", "pages", "quizzes"}

\item{include}{optional additions to the query string}
}
\value{
data.frame
}
\description{
Function to return various course items. See "item" argument below. Omitting the "item argument
returns a course object.
}
\examples{
#' get_course_items(course_id = 20, item = "settings")
#' get_course_items(course_id = 20, item = "enrollments")
}

