% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-data.R
\name{get_course_members}
\alias{get_course_members}
\title{Function to return course members.}
\usage{
get_course_members(course_id, type = "students")
}
\arguments{
\item{course_id}{valid Canvas course id}

\item{type}{one of "students" or "users"}
}
\value{
data.frame
}
\description{
Function to return course members.
}
\examples{
#' get_course_members(course_id = 20)
#' get_course_members(course_id = 20, type = "users")
}

