% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discussions.R,
%   R/get_discussion_entries_by_user.R
\name{get_discussion_entries_by_user}
\alias{get_discussion_entries_by_user}
\title{Get discussion entries by user}
\usage{
get_discussion_entries_by_user(course_id, user_id)

get_discussion_entries_by_user(course_id, user_id)
}
\arguments{
\item{course_id}{The course ID.}

\item{user_id}{The user ID.}
}
\value{
A data frame containing discussion entries for the specified user.

A data frame containing discussion entries for the specified user.
}
\description{
Get discussion entries by user

Get discussion entries by user
}
