% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_folder_list}
\alias{get_folder_list}
\title{get folder information}
\usage{
get_folder_list(id, id_type, content_types = NULL, search_term = NULL,
  include = "user", only = NULL, sort = "name", order = "asc")
}
\arguments{
\item{id}{can be one of course, section, or user id.  Can be input as numbers or characters. If a user has multiple enrollments in a context (e.g. as a teacher and a student or in multiple course sections), each enrollment will be listed separately.}

\item{id_type}{the type id being used.  Can be 'courses', 'sections', or 'users'. Defaults to 'courses'.}

\item{content_types}{a    string. Filter results by content-type. You can specify type/subtype pairs (e.g., 'image/jpeg'), or simply types (e.g., 'image', which will match 'image/gif', 'image/jpeg', etc.).}

\item{search_term}{a string. The partial name of the files to match and return.}

\item{include}{a string. Array of additional information to include. “user” the user who uploaded the file or last edited its content. “usage_rights” copyright and license information for the file (see UsageRights). Allowed values: user}

\item{only}{an Array. Array of information to restrict to. Overrides include. “names” only returns file name information.}

\item{sort}{a string. Sort results by this field. Defaults to 'name'. Note that `sort=user` implies `include[]=user`. Allowed values:   name, size, created_at, updated_at, content_type, user}

\item{order}{a    string. The sorting order. Defaults to 'asc'. Allowed values: asc, desc.}
}
\description{
This function pull enrollment information. Depending on the `id_type` given, return a paginated list of either (1) all of the enrollments in a course, (2) all of the enrollments in a section or (3) all of a user's enrollments. This includes student, teacher, TA, and observer enrollments. note: Currently, only a root level admin user can return other users' enrollments. A user can, however, return his/her own enrollments.
}
