% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{get_group_categories}
\alias{get_group_categories}
\title{Group categories}
\usage{
get_group_categories(course_id)
}
\arguments{
\item{context_type}{}
}
\description{
Group categories
}
\examples{
get_group_categories(1350207)
}
