% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{get_group_category}
\alias{get_group_category}
\title{Get a single group category}
\usage{
get_group_category(group_category_id)
}
\arguments{
\item{group_category_id}{id number for the category}
}
\value{
dataframe of category ids
}
\description{
Get a single group category
}
\examples{
\dontrun{get_group_category(52872)}
}
