% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{get_group_users}
\alias{get_group_users}
\title{Get group users}
\usage{
get_group_users(group_id, group_name)
}
\arguments{
\item{group_id}{which group}
}
\value{
users in a group
}
\description{
Get group users
}
