% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizzes.R
\name{get_quiz}
\alias{get_quiz}
\title{Get a single Quiz object}
\usage{
get_quiz(course_id, quiz_id)
}
\arguments{
\item{course_id}{Course ID}

\item{quiz_id}{Quiz ID}
}
\value{
Data frame representation of the Quiz object
}
\description{
Get a single Quiz object
}
