% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizzes.R
\name{get_quiz_list}
\alias{get_quiz_list}
\title{Function to list all quizzes}
\usage{
get_quiz_list(course_id = NULL)
}
\arguments{
\item{course_id}{Course ID}
}
\value{
data frame
}
\description{
Function to list all quizzes
}
\examples{
#' get_quiz_list()
}
