% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_response.R
\name{process_response}
\alias{process_response}
\title{Process a Canvas API response

Wrapper function for common tasks in going from Canvas url to dataframe. Most
of the heavy lifting is done in \code{get_pages}, which finds which pages to
download. This function adds necessary arguments to those pages (e.g. the
authentication token), downloads the content, converts from JSON into data
frame format, and if there are multiple pages/dataframes, converts it into
one final dataframe.}
\usage{
process_response(url, args)
}
\arguments{
\item{url}{url to query}

\item{args}{query arguments to be passed to \code{httr}, e.g. auth token}
}
\value{
processed dataframe
}

